amp_simulated_servers/requirements.txt
appnope==0.1.0
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
mistune==0.8.4
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.6.0
prometheus-flask-exporter==0.5.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.0
requests==2.21.0
ruamel.yaml==0.15.89
Send2Trash==1.5.0
six==1.12.0
urllib3==1.24.1
uWSGI==2.0.18
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
